requirements.txt
asgiref==3.9.1
boto3==1.39.3
botocore==1.39.3
dj-database-url==2.1.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
django-storages boto3
boto3
